FIX-506: move header
