+++ include/eve/detail/compress/simd/common/compress_using_switch.hpp
@@ -13,7 +13,7 @@
 #include <eve/function/unalign.hpp>
 #include <eve/function/shuffle.hpp>
 
-#include <eve/detail/top_bits.hpp>
+#include <eve/arch/top_bits.hpp>
 
 namespace eve::detail
 {
