+++ include/eve/detail/compress/simd/x86/compress_using_bmi.hpp
@@ -12,7 +12,7 @@
 #include <eve/function/if_else.hpp>
 #include <eve/function/slide_left.hpp>
 
-#include <eve/detail/top_bits.hpp>
+#include <eve/arch/top_bits.hpp>
 
 namespace eve::detail
 {
