+++ include/eve/module/core/regular/impl/count_true.hpp
@@ -10,7 +10,7 @@
 #include <eve/concept/value.hpp>
 #include <eve/concept/conditional.hpp>
 #include <eve/detail/implementation.hpp>
-#include <eve/detail/top_bits.hpp>
+#include <eve/arch/top_bits.hpp>
 
 namespace eve::detail
 {
