+++ include/eve/module/core/regular/impl/first_true.hpp
@@ -10,7 +10,7 @@
 #include <eve/concept/value.hpp>
 #include <eve/logical.hpp>
 #include <eve/detail/has_abi.hpp>
-#include <eve/detail/top_bits.hpp>
+#include <eve/arch/top_bits.hpp>
 #include <eve/detail/implementation.hpp>
 #include <eve/function/any.hpp>
