+++ include/eve/module/core/regular/impl/if_else.hpp
@@ -8,7 +8,6 @@
 #pragma once
 
 #include <eve/detail/implementation.hpp>
-#include <eve/detail/top_bits.hpp>
 #include <eve/concept/compatible.hpp>
 #include <eve/concept/generator.hpp>
 #include <eve/concept/value.hpp>
