+++ include/eve/module/core/regular/impl/simd/arm/neon/first_true.hpp
@@ -7,7 +7,7 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/detail/top_bits.hpp>
+#include <eve/arch/top_bits.hpp>
 #include <eve/function/any.hpp>
 #include <eve/function/convert.hpp>
